+++ test/content-routing/dht/utils.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const KadDht = require('libp2p-kad-dht')
-const { multicodec } = require('libp2p-kad-dht')
 const Crypto = require('../../../src/insecure/plaintext')
 const Muxer = require('libp2p-mplex')
 const Transport = require('libp2p-tcp')
@@ -25,13 +24,12 @@ const subsystemOptions = mergeOptions(baseOptions, {
   config: {
     dht: {
       kBucketSize: 20,
-      randomWalk: {
-        enabled: true
-      },
       enabled: true
     }
   }
 })
 
 module.exports.subsystemOptions = subsystemOptions
-module.exports.subsystemMulticodecs = [multicodec]
+module.exports.subsystemMulticodecs = [
+  '/ipfs/lan/kad/1.0.0'
+]